backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
Markdown==3.10.2
packaging==25.0
python-dotenv==1.2.2
PyYAML==6.0.3
sqlparse==0.5.3
